+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v0.6.3
+
+* [FIXED] `serviceIntegrity.inspectCouchbase` - if a bucket is not connected already, wait on `connect` or `error` event before the test get request is made
+* [FIXED] don't print warning to stdout when `SEQEULIZE_DEBUG` is `undefined`
+* [FIXED] `moduleLoader` - don't load unsupported files... load only files with an extension from `require.extensions`
+
 ## v0.6.2
 
 * [FIXED] `staticData.loadSync` - convert stdout, stderr Buffers to strings (incorrect condition caused the buffers to NOT be converted)